+++ core/dhara_storage_core/src/types/mod.rs
@@ -9,6 +9,9 @@ mod permissions;
 mod size;
 
 pub use attributes::StorageAttributes;
-pub use options::{DirectoryDeleteOptions, ReadOptions, TransferOptions, WriteOptions};
+pub use options::{
+    DirectoryDeleteOptions, FileShareMode, OpenReadOptions, OpenWriteOptions, ReadOptions,
+    TransferOptions, WriteOptions,
+};
 pub use permissions::StoragePermissions;
 pub use size::{SizeUnit, StorageSize, format_size};